new file:   tests/relax_test/Session.vim
Added "-print_uniq" to see the solutions before removing smaller ones in final output.

Currently lot's of results are begin generated even for very simple examples.
It may be that the example I use has too little variation though.

The example is in:
tests/relax_test

The idea with the example is that in all files there is a
f(X)->f(X,X)
change and in one file there is a
42 -> 42+42+42
change. The example will be updated to reflect that in the next commit.