modified:   main.ml
We are now using the real datamining to filter out the "omnipresent" patches of
the initial phase. This gives a *SIGNIFICANT* speedup.

modified:   visitor_j.ml

Additional constructs handled in the translation. This has an impact on the
size of the gterms being generated and thusly one should increase the value of
the "-top" parameter to spdiff.

modified:   tests/linux-seq-path/specfile

We can now handle all the files in this directory due to the changes in the
translation (visitor).