Inspecting whether using a more clever way to do the apply function have impact on the running time.
When applying p->p' to some t, we first look whether p matches some
subterm of t before trying to apply the patch. The motivation is that if
for some large subterm, p does not match we do not need to rebuild that
subterm (as part of the traversal of it) but can use the one we already
have.

I am not completely sure, that the apply function does not already do
that though!