Better pattern inference support
In this commit the example (paths-ex) works as expected.

The idea is to first find a set of node-patterns that are common in all graphs. Then, we doing
the pattern growing consider only the nodes of the graph in question that unfify with some common
node-pattern as just found.

For the example, it gives no noticable speed-up, but it should allow us to not grow from nodes
that have nothing in common with the other graphs.