modified:   diff.ml modified:   main.ml
Introduced two new abstraction reduction measurements:
- only allow extending a spat with a new pattern if the pattern either
	has no metas or shares at least one meta with the spat to be extended
- from abstract_term in grow, apply filter_less_abstract so that only
	the most abstract abstractions of a term is used

These two have "interesting" interaction when applied to a dataset. If
the abstractness is kept low, one might find a common pattern because we
return less abstract abstractions so we have more chance to extend,
because we have no "sharing constraint" from the first measurement
above. In the end, the result is that increasing the abstractness factor
can make the program return *fewer* (or none) results.

This might be changed in the future!