modified:   diff.ml
Introduced a new cont_match_spec function which takes a "spec" function
to apply for concrete matches. The spec is applied to the "CM pattern"
and the term of the node. It should return a binding of metavariables in
the pattern. It can raise ErrorSpec if it finds out that the spec should
under no circumstances hold. Kind of a hack used in the new
"find_embedded_succ". find_embedded_succ g n p starts from node n and
checks whether pattern p occurs (nested) in all subsequent paths.

modified:   main.ml

Changed the abstract_term function so that function identifiers are
*never* abstracted. Also, always abstract a subterm if for some location
where the top-level term matches on all paths from that location the
subterm always eventually occur (nested).

modified:   visitor_j.ml

Translation modified so that the assignment operator is part of the
exp-type instead of a separate entity in the gtree. Thus, we never
abstract the assignment operator with a metavariable. At least the
ase-ex1 example still works out nicely for generic patch inference.