modified:   ../../diff.ml modified:   ../../main.ml
First support for finding nested patterns. Watch out for how to use.
Next step is to consider if the top-param can be taken out or if there
is a sane way to reduce the number of abstractions tried. That is, for
some reason it now thinks X0 is sane to try.

In the end we really wish to get rid of "abstractness" and "top"
parameters but maybe give some option as to what is sane?