More "..." work:
Removed "-top" parameter (or made it defunct). It has always annoyed me that
one had to specify this. Instead I plan to just not abstract in terms that is
already know to be too large. So far only C("def", _) is considered too large.

There is still a bug in locate_subterm that needs to be found.

Also, eventually I will need to be able to construct CFGs from gtrees and that
will make the complicated 'locate_subterm' redundant.