modified:   diff.ml
When we compute δ(t,t')+δ(t',t'') = δ(t,t'') and it turns out that
δ(t,t')+δ(t',t'') < δ(t,t'') we would previously raise an exception and the
program would stop running. Now we print some info on the output and let the
function return false so that we continue computation.

At some point I will have to figure out why it can return "<"!