modified:   main.ml
Change in locate_subterm, when we reach a singleton-path and the node-term is
not the point of interest (context_point) of the current subterm should we bail
out completely or should we raise the LocateSubterm exception to allow the
search for the context-point to contine on another branch of the subterm?

This commit changes the behaviour into the latter under the assumption that we
might end in wrong subterms after a goto -- in which case we call locate_label
that does not know about when a gtree is a statement or not causing the
locate_subterm (loop) to be called on those branches.

A better fix would only iterate over statements and nested-statements.

modified:   visitor_j.ml

Handle arguments to macroiterators.