modified:   diff.ml
Fixed find_simple_updates_merge_changeset so that 2 * 10^18 possible
combinations need not all be examined.

The problem was:
1) When we add a solution to the accumulated ones we (now) check if it
is already there by structural equality. Thus, it becomes important to
normalize the use of metavariables.

This further required a fix in 'renumber_metas_up' to not use the
assertion.

2) When we find the non-abtract updates with respect to a single pair we
should remove duplicates and also we should only return the *smallest*
updates since ... PENDING argument.