Fix error-exits
When we encounter an error-exit path, we raise the Bailout exception
and when we catch it we replace it with "true". However, if there only
path leads to an error-exit, we can therefore have any suffix attached
to patterns that end in a node which only has error-exits as
successor.

The fix changes the behaviour so that one matching the successors in
"..." clause we require that at least one is not an error-exit. There
could still be an issue if a node has a self loop and otherwise an
error-exit because we treat loops as "true" under the assumption that
eventually some other path leads to a result.