Fix a bug in application of semantic patches
Previously a patch such as

  f();
  ...
  g();
  ...
+ add();
  return 42;

would be considered safe for the following pair of terms:

original function:

void foo() {
	f();
	if(b) {
	  g();
	}
	return 42;
}

modified function:

void foo() {
	f();
	if(b) {
  	g();
	  add();
	}
	return 42;
}

This is very strange because it should not be safe to add the
function-call to 'add' just before the return-statement.

The problem was actually to be found in the function for applying
semantic (not just generic) patches.