Fixed bug with loops in spattern matching
There was a bug where looping constructs such as
f();
for(;;)
  g();
h();

would completely fool the pattern-matcher and have any pattern with a
prefix

f();
...
g();
...

match the above program and result in an empty matching-trace. The
consequence was that application of any spatch based on the semantic
pattern would do no change (i.e. be the identity transformation) to
the file.

Thus, any transformation based on the spattern would be considered
safe!