Fixed another bug in matcher + perfo. optimization
There was a bug in the function matching semantic patterns that would
allow the matcher to skip the same node with "..." multiple times and
thus effectively entering an infinite loop.

Also, implemented a potential performance optimization:
If a focus_function is given, the matcher tries to match candiate
patterns with the focus_function first. Only if the pattern matches
the focus_function, will it try to match the pattern with other
functions.

Thus: in case a pattern matches at least 'threshold' other functions
but does not match the focus_function, the mentioned change is much
faster. In all other cases, performance is the same.