Re-implement matching of semantic patterns
The previous approach for matching a semantic pattern (a ... b) used
a naive CPS approach in which it would frequently visit the same nodes
several times.

For CFGs with switches, loops, and gotos the performance of the CPS
matcher would be really bad.

The new approach works much more directly:

To test  n |- a ... b ... c we first check n |- a and then use a
depth-first strategy to find all subsequent nodes where b occurs.
At those b-nodes (nb) we then try test nb |- b ... c and use this
result to match with a.
perform