Bug-fix for modification of labeled terms
When applying a patch
- p
to a statement
lab: s

we would be able to figure out that p
could match s and thus be removed resulting in
lab:

but: the internal representation would say that
there was an empty statement labeled with 'lab'

This is a problem if we have the original code:

lab: a;
     b;

and patch
- a;

and it was supposed to have been:

lab: b;

The spatch_apply_fixed function would apply
the "-a" patch seemingly correct, but in fact
it should have inlined the "b;" statement so
that it would now be labeled with 'lab'.

To solve the problem, we check for such empty
labeled statements and inline the subsequent one.

There's a potential problem if the empty-labeled
statement is the last in a compound such as:

{
     b;
lab: a;
}