First steps towards making spdiff faster
Basically the idea is that we should find patterns based on chunks
that we detected as having changed. Now, we find patterns from a
premade set of node patterns.

We should still construct the set of initial node patterns, but when
deciding which node pattern to start growing a larger pattern from, we
should only consider the node patterns that matched a changed
location.

If we implement the above, we have to take care to grow patterns in
"both" directions because otherwise there might be some patterns that
starting with an unchanged node-pattern that we will not find.