Make part of spdiff use more cores
This commit requires that the Parmap library gets installed
(https://gitorious.org/parmap) the makefile assumes that the library
is installed in /usr/lib/ocaml/parmap.

Also required is the extlib (http://code.google.com/p/ocaml-extlib/)
assumed to be in /usr/lib/ocaml/extlib.

The location can be changed with the SYSDIR variable in the Makefile.