Fix use of parmap library
I forgot that when 'concat'ing several lists obtained from seperate
cores, we also need to merge those lists.

The result was that when the number of input term-pairs was exactly 4
(or maybe even factor of 4)--because we make parmap use 4 separate
threads-- we would not actually do any merging of the lists.