build: add rules for .ml generation from .mly and .mll
The build here did not finish because here because it was missing
- genericlexer.ml
- genericparser.ml
- genericparser.mli
- genericparser.cmi

as input files. This files are autogenerated. This patch two make rules
so they can be autocreated during the build. Additionally the
genericparser.mli file has been added to clean up rules as it is also
auto generated.
With this change was able to build it from scratch for all and opt
target.

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>